fix: align frontend API base URL defaults
