+++ .env.example
@@ -11,4 +11,4 @@ PUBLIC_CHAT_API_KEY=
 # Example environment variables for local development
 # Base path for the backend API used by the SvelteKit frontend
 # e.g. /api or https://your-deploy.com/api
-PUBLIC_API_BASE_URL=http://localhost:8080/api
+PUBLIC_API_BASE_URL=/api
