+++ backend/tests/test_body_size_limit.py
@@ -3,20 +3,7 @@ import sys
 import asyncio
 import httpx
 from pathlib import Path
-import types
 import pytest
-
-# stub redis module for tests
-redis_stub = types.ModuleType("redis")
-redis_asyncio_stub = types.ModuleType("redis.asyncio")
-
-def from_url(*args, **kwargs):
-  return None
-
-redis_asyncio_stub.from_url = from_url
-redis_stub.asyncio = redis_asyncio_stub
-sys.modules["redis"] = redis_stub
-sys.modules["redis.asyncio"] = redis_asyncio_stub
 
 os.environ["OPENAI_API_KEY"] = "test"
 os.environ["CHAT_API_KEY"] = "test-key"
