+++ backend/tests/test_template_checksums.py
@@ -1,20 +1,7 @@
 import os
 import sys
-import types
 from pathlib import Path
 import hashlib
-
-# stub redis module for import side effects
-redis_stub = types.ModuleType("redis")
-redis_asyncio_stub = types.ModuleType("redis.asyncio")
-
-def from_url(*args, **kwargs):
-  return None
-
-redis_asyncio_stub.from_url = from_url
-redis_stub.asyncio = redis_asyncio_stub
-sys.modules["redis"] = redis_stub
-sys.modules["redis.asyncio"] = redis_asyncio_stub
 
 os.environ["OPENAI_API_KEY"] = "test"
