+++ frontend/src/routes/+page.svelte
@@ -19,7 +19,7 @@
     input = ''
 
     // 3️⃣ Call your backend
-    const baseUrl = sanitizeBaseUrl(import.meta.env.PUBLIC_API_BASE_URL || '/api')
+    const baseUrl = sanitizeBaseUrl(import.meta.env.PUBLIC_API_BASE_URL ?? '/api')
     const url = `${baseUrl}/chat`
     try {
       const apiKey = import.meta.env.PUBLIC_CHAT_API_KEY
